requirements.txt
langchain
langchain-community
langchain-openai
openai
faiss-cpu
tiktoken
pypdf
rank_bm25
streamlit
python-dotenv
